pyproject.toml
[tool.poetry]
name = "django-tutorial"
version = "0.1.0"
description = ""
authors = ["ukastsiukevich <grun_gespenst@tut.by>"]
readme = "README.md"
packages = [{include = "django_tutorial"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.1.3"
django-debug-toolbar = "^4.4.6"
flake8 = "^7.1.1"
python-dotenv = "^1.0.1"
isort = "^5.13.2"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
ruff = "^0.8.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "test.settings"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
